Template for rflood
I've made a template to use rflood as an alternative for qBittorrent since the UI is a bit more cleaner. I was surprised at how little extra configuration is needed to get it all running, however finding the sock file was an issue but eventually found it in `/dev/shm/rtorrent.sock` I've never worked with sock files before so never knew where they were at.

Hopefully it'll be some use for some!